Enable Windows python bindings (#343)
* Use int64 for counter to fix windows compilation error

* Fix windows python bindings by adding install_lib command to move windows build output into python package

* Update to use Path instead of os

* Change batch_insert num_inserts signature to signed type for OpenMP compatibility

* Update num_inserts to int32_t per PR request

---------

+++ python/src/diskann_bindings.cpp
@@ -194,7 +194,7 @@ template <class T> struct DynamicInMemIndex
     }
 
     auto batch_insert(py::array_t<T, py::array::c_style | py::array::forcecast> &vectors,
-                      py::array_t<IdT, py::array::c_style | py::array::forcecast> &ids, const size_t num_inserts,
+                      py::array_t<IdT, py::array::c_style | py::array::forcecast> &ids, const int32_t num_inserts,
                       const int num_threads = 0)
     {
         if (num_threads == 0)
@@ -204,7 +204,7 @@ template <class T> struct DynamicInMemIndex
         py::array_t<int> insert_retvals(num_inserts);
 
 #pragma omp parallel for schedule(dynamic, 1)
-        for (size_t i = 0; i < num_inserts; i++)
+        for (int32_t i = 0; i < num_inserts; i++)
         {
             insert_retvals.mutable_data()[i] = _index->insert_point(vectors.data(i), *(ids.data(i)));
         }
